refactor: move listening logic from profile core
